메세지 전송 상태 UI 수정
   - JND-2756

   메세지 내에 전송 상태 아이콘 추가

+++ app/src/main/java/com/tosslab/jandi/app/ui/message/v2/loader/NormalNewMessageLoader.java
@@ -79,7 +79,8 @@ public class NormalNewMessageLoader implements NewsMessageLoader {
                 .observeOn(Schedulers.io())
                 .subscribe(message -> {
                     message.roomId = roomId;
-                    if (TextUtils.equals(message.message.status, "archived")
+                    if (!TextUtils.equals(message.status, "event")
+                            && TextUtils.equals(message.message.status, "archived")
                             && !(message.message instanceof ResMessages.FileMessage)) {
                         MessageRepository.getRepository().deleteMessage(message.messageId);
                     } else {
